NahaXMLUtils

Documentation for NahaXMLUtils.

NahaXMLUtils.bounds_to_viewbox

NahaXMLUtils.elt

NahaXMLUtils.svgval

NahaXMLUtils.bounds_to_viewbox — Method

bounds_to_viewbox(::Bounds)

Returns a vector of Pairs which provide the viewBox, width and height attributes to elt for an svg element.

NahaXMLUtils.elt — Method

elt(f, tagname::AbstractString, things...)
elt(tagname::AbstractString, things...)

Return an XML element. f is called with a single argument: either an XML.AbstractXMLNode or a Pair describing an XML attribute to be added to the resulting element.

Each of things is either an XML.AbstractXMLNode or a Pair describing an XML attribute to be added to the resulting element.

Each of things, and whatever f is called on are added to the new XML element.

An argument could also be either a string or a number, in which case it is included as text content.

If an argument is an AbstractDict then each entry is added as an attribute.

NahaXMLUtils.svgval — Function

svgval(value)

return a representation of value that can appear in SVG code.